==> real_estate/v3/01_create_dataframe.ipynb <==
import pandas as pd

dataframes = {
    "01": pd.read_csv("~/real_estate/data/raw/Zip_MedianValuePerSqft_AllHomes.csv", encoding='ISO-8859-1'),
    "02": pd.read_csv("~/real_estate/data/raw/Zip_PctOfHomesDecreasingInValues_AllHomes.csv", encoding='ISO-8859-1'),
    "03": pd.read_csv("~/real_estate/data/raw/Zip_PctOfHomesIncreasingInValues_AllHomes.csv", encoding='ISO-8859-1'),
    "04": pd.read_csv("~/real_estate/data/raw/Zip_Zhvi_1bedroom.csv", encoding='ISO-8859-1'),
    "05": pd.read_csv("~/real_estate/data/raw/Zip_Zhvi_2bedroom.csv", encoding='ISO-8859-1'),
    "06": pd.read_csv("~/real_estate/data/raw/Zip_Zhvi_3bedroom.csv", encoding='ISO-8859-1'),
    "07": pd.read_csv("~/real_estate/data/raw/Zip_ZriPerSqft_AllHomes.csv", encoding='ISO-8859-1'),
    "08": pd.read_csv("~/real_estate/data/raw/Zip_Zhvi_Condominum.csv", encoding='ISO-8859-1'),
    "09": pd.read_csv("~/real_estate/data/raw/Zip_Zhvi_SingleFamilyResidence.csv", encoding='ISO-8859-1'),
    "10": pd.read_csv("~/real_estate/data/raw/fred_3monthtreasurybill.csv"),
    "11": pd.read_csv("~/real_estate/data/raw/fred_average_hourly_earnings_seasonally_adjusted.csv"),
    "12": pd.read_csv("~/real_estate/data/raw/fred_fixed_mortgage_rate.csv"),
    "13": pd.read_csv("~/real_estate/data/raw/fred_total_nonfarm_payroll.csv"),
    "14": pd.read_csv("~/real_estate/data/backup/irs_zipcode_data.csv")
}

zillow_variables = {
    "01": "median_sqft_value",
    "02": "percent_decreasing",
    "03": "percent_increasing",
    "04": "zhvi_1bed",
    "05": "zhvi_2bed",
    "06": "zhvi_3bed",
    "07": "zri_sqft_value",
    "08": "zhvi_condo",
    "09": "zhvi_singlefam"
}

fred_variables = {
    "10": "90_day_treasury_bill_rate",
    "11": "mean_hourly_earnings_adj",
    "12": "mean_fixed_mortgage_rate",
    "13": "total_nonfarm_payroll"
}

irs_variables = {
    "14": "people_count",
    "15": "adult_count",
    "16": "mean_adjusted_income",
    "17": "mean_adjusted_income_with_dep",
    "18": "mean_real_estate_taxes_hh",
    "19": "mean_mortgage_interest_hh"
}

dataframes_melt = {}

for i in zillow_variables.keys():
    
    df_ = dataframes[i].drop(["RegionID","SizeRank","City","State","Metro","CountyName","SizeRank"],axis=1)
    dataframes_melt[i] = pd.melt(df_, id_vars=["RegionName"]).rename(columns={"variable":"date",
                                                                                "value":zillow_variables[i]})

for i in fred_variables.keys():
    
    varname = dataframes[i].columns[1]
    df_ = dataframes[i].rename(columns={"DATE":"date",varname:fred_variables[i]})
    df_["date"] = [j[:7] for j in df_["date"].values]
    df_ = df_.groupby(["date"], as_index=False).agg('mean')
    dataframes_melt[i] = df_

# Join all zillow dfs
df_zillow = pd.merge(dataframes_melt["01"], dataframes_melt["02"], on=["RegionName","date"], how="outer")
for i in ["03","04","05","06","07","08","09"]:
    df_zillow = pd.merge(df_zillow, dataframes_melt[i], on=["RegionName","date"], how="outer")
    
# Merge with Fed dfs
df = pd.merge(df_zillow, dataframes_melt["10"], on=["date"], how="left")
for i in ["11","12","13"]:
    df = pd.merge(df, dataframes_melt[i], on=["date"], how="left")

# Filter dates, remove NaNs
df = df.loc[df["date"] > '2000']
df = df.dropna()

# Date format
df["month"] = [int(i[5:]) for i in df["date"].values]
df["year"] = [int(i[:4]) for i in df["date"].values]
df = df.drop(["date"], axis=1)
    
# Merge with IRS dfs
dataframes["14"] = dataframes["14"][["zipcode","people_count_lag","adult_count_lag","mean_adjusted_income_lag",
                                     "mean_adjusted_income_with_dep_lag","mean_real_estate_taxes_hh_lag",
                                     "mean_mortgage_interest_hh_lag","year","state"]]
dataframes["14"] = dataframes["14"].rename(columns={"zipcode":"RegionName"})
df = pd.merge(df, dataframes["14"], on=["RegionName","year"], how="left")
df = df.dropna()

# Dummies on state
df = pd.concat([df, pd.get_dummies(df["state"])], axis=1)
df = df.drop(["state"], axis=1)

df.head()

df.to_csv("~/real_estate/data/v3/unsupervised_df.csv",index=False)
