==> real_estate/v3/05_pca.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.decomposition import PCA
from sklearn.externals import joblib

%matplotlib inline
np.random.seed(42)

# Load data
train_df = pd.read_csv("~/real_estate/data/v3/train_df.csv")
test_df = pd.read_csv("~/real_estate/data/v3/test_df.csv")

features = train_df.columns[:-2]

X_train = train_df[features]
y_train = train_df["target"]
X_test = test_df[features]
y_test = test_df["target"]

### PCA

# Fit PCA
pca = PCA(n_components=45)
pca.fit(X_train.values)
print("Total explained variance: ", sum(pca.explained_variance_ratio_))

# Reduce dimensionality
train_df_pca = pca.transform(X_train.values)
train_df_pca = pd.DataFrame(train_df_pca)
train_df_pca["target"] = y_train.values

test_df_pca = pca.transform(X_test.values)
test_df_pca = pd.DataFrame(test_df_pca)
test_df_pca["target"] = y_test.values
test_df_pca["RegionName"] = test_df["RegionName"].values

train_df_pca.head()

train_df_pca.to_csv("~/real_estate/data/v3/train_df_pca.csv",index=False)
test_df_pca.to_csv("~/real_estate/data/v3/test_df_pca.csv",index=False)
joblib.dump(pca, "/home/gnazareths/real_estate/data/v3/pca.pkl")
